Convert the ICal_Generator class into Namespaced functions.
